requirements-prd.txt
asgiref==3.6.0
boto3==1.34.11
botocore==1.34.11
cffi==1.15.1
cryptography==40.0.2
Django==4.2
django-filter==23.3
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-storages==1.14.2
djangorestframework==3.14.0
jmespath==1.0.1
jsonfield==3.1.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
urllib3==2.0.7
python-memcached==1.62
pymemcache==4.0.0
redis==5.0.1
